db/requirements.txt
psycopg2==2.9.2
Flask==2.0.2
python-dotenv==0.19.2
openpyxl==3.0.9
